imgproc: contours C-API cleanup
